Now adapt the ~/Sites/PlanScore/planscore/tests/test_score.py test_vectorized_swing to this new function and ensure the new test passes
Add comprehensive test_vectorized_logit_shift test

- Created test_vectorized_logit_shift() modeled after test_vectorized_swing()
- Tests zero shift, positive shift, negative shift scenarios
- Verifies national vote share changes correctly (not uniform per-district)
- Tests competitive districts shift more than safe districts (logit property)
- Tests zero-vote district handling
- Tests higher-dimensional arrays (5D: model_versions × incumbency × sims × districts × parties)
- Fixed bug in vectorized_logit_shift(): use dtype=float for new_votes to preserve turnout exactly
- All 195 tests pass (194 existing + 1 new)
- Test runtime: 6.759s

+++ planscore/score.py
@@ -334,7 +334,7 @@ def vectorized_logit_shift(votes:numpy.typing.NDArray, target_diff:float) -> num
 
     # Convert back to vote counts
     # Shape: (N, districts, 2)
-    new_votes = numpy.zeros_like(votes_flat)
+    new_votes = numpy.zeros_like(votes_flat, dtype=float)
     new_votes[:, :, 0] = new_shares * turnout  # blue
     new_votes[:, :, 1] = (1 - new_shares) * turnout  # red
